full conversion to t_size, started degenerate zero case

+++ src/model.h
@@ -15,7 +15,7 @@ class model
 {
   double lambda; // penalty
 protected:
-  int df; // degrees of freedom
+  size_t df; // degrees of freedom
 public:
   model(void);
   model(double);
@@ -23,7 +23,7 @@ public:
   virtual double mutual_information(variable, variable) = 0;
   double aic(variable, variable); // an information criterium
   double penalized_information(variable, variable);
-  int get_df() const;
+  size_t get_df() const;
 };
 
 #endif
